Cluster: units to float

+++ api/models/Cluster/Report/Beneficiaries.js
@@ -206,7 +206,7 @@ module.exports = {
 		},
 			// units -> eiewg
 		units: {
-			type: 'integer',
+			type: 'float',
 			defaultsTo: 0
 		},
 		// units
